+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/item/MessageTextItem.kt
@@ -31,6 +31,7 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import me.saket.bettermovementmethod.BetterLinkMovementMethod
+import java.net.URL
 
 @EpoxyModelClass(layout = R.layout.item_timeline_event_base)
 abstract class MessageTextItem : AbsMessageItem<MessageTextItem.Holder>() {
@@ -49,15 +50,25 @@ abstract class MessageTextItem : AbsMessageItem<MessageTextItem.Holder>() {
     private val mvmtMethod = BetterLinkMovementMethod.newInstance().also {
         it.setOnLinkClickListener { _, url ->
             // Return false to let android manage the click on the link, or true if the link is handled by the application
-            urlClickCallback?.onUrlClicked(url) == true
+            try {
+                (URL(url))
+                urlClickCallback?.onUrlClicked(url) == true
+            } catch (t: Throwable) {
+                false
+            }
         }
         // We need also to fix the case when long click on link will trigger long click on cell
         it.setOnLinkLongClickListener { tv, url ->
             // Long clicks are handled by parent, return true to block android to do something with url
-            if (urlClickCallback?.onUrlLongClicked(url) == true) {
-                tv.dispatchTouchEvent(MotionEvent.obtain(0, 0, MotionEvent.ACTION_CANCEL, 0f, 0f, 0))
-                true
-            } else {
+            try {
+                (URL(url))
+                if (urlClickCallback?.onUrlLongClicked(url) == true) {
+                    tv.dispatchTouchEvent(MotionEvent.obtain(0, 0, MotionEvent.ACTION_CANCEL, 0f, 0f, 0))
+                    true
+                } else {
+                    false
+                }
+            } catch (t: Throwable) {
                 false
             }
         }